JSDK-263 【点播】根据分类批量获取视频时长和大小，修改文件大小字段类型

+++ polyv-java-vod-sdk/src/main/java/net/polyv/vod/v1/entity/manage/info/VodGetVideoSizeResponse.java
@@ -53,19 +53,19 @@ public class VodGetVideoSizeResponse {
          * 编码后码率1FLV的大小
          */
         @ApiModelProperty(name = "filesize1", value = "编码后码率1FLV的大小")
-        private Integer filesize1;
+        private Long filesize1;
         
         /**
          * 编码后码率2FLV的大小
          */
         @ApiModelProperty(name = "filesize2", value = "编码后码率2FLV的大小")
-        private Integer filesize2;
+        private Long filesize2;
         
         /**
          * 编码后码率3FLV的大小
          */
         @ApiModelProperty(name = "filesize3", value = "编码后码率3FLV的大小")
-        private Integer filesize3;
+        private Long filesize3;
     }
     
 }
